Rename j-placeholder.js -> JPlaceholder.js
