+++ elements.js
@@ -7,7 +7,7 @@ import {JDialog} from './src/components/j-dialog.js';
 import {JField} from './src/components/JField.js';
 import {JIcon} from './src/components/JIcon.js';
 import {JInput} from './src/components/JInput.js';
-import {JPlaceholder} from './src/components/j-placeholder.js';
+import {JPlaceholder} from './src/components/JPlaceholder.js';
 import {JTooltip} from './src/components/j-tooltip.js';
 
 // Utilities
